claude: refactor: move message helper functions to dym_kas_core::message module
These functions operate on HyperlaneMessage types and don't use any provider state, so they belong in the message utilities module, not as methods on KaspaProvider.
